dev_requirements.txt
nose==1.3.6
django-nose==1.4
pylint==1.4.3
flake8==2.4.0
factory-boy==2.5.2
python-memcached==1.54
